Add code from manual

+++ MetricsAgent/DAL/Handlers/DateTimeOffsetHandler.cs
@@ -7,7 +7,7 @@ namespace MetricsAgent.DAL.Handlers
     public class DateTimeOffsetHandler : SqlMapper.TypeHandler<DateTimeOffset>
     {
         public override DateTimeOffset Parse(object value)
-            => DateTimeOffset.FromUnixTimeSeconds((int)value);
+            => DateTimeOffset.FromUnixTimeSeconds((long)value);
 
         public override void SetValue(IDbDataParameter parameter, DateTimeOffset value)
         {
